# docs/swagger.yaml
definitions:
  syncer.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: 区块链浏览器后端API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 区块链浏览器
  version: "1.0"
paths:
  /addresses:
    get:
      consumes:
      - application/json
      description: 分页获取账户地址列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 账户地址列表
      tags:
      - 账户地址
...
  /assets:
    get:
      consumes:
      - application/json
      description: 分页获取资产列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 资产列表
      tags:
      - 资产
...
  /blocks:
    get:
      consumes:
      - application/json
      description: 分页获取区块列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 区块列表
      tags:
      - 区块
  /blocks/{name}:
    get:
      consumes:
      - application/json
      description: 获取区块信息
      parameters:
      - description: 区块高度/区块哈希
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 区块信息
      tags:
      - 区块
...
  /contractcodes:
    get:
      consumes:
      - application/json
      description: 分页获取合约代码列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 合约代码列表
      tags:
      - 合约代码
  /contractcodes/{name}/transactions:
    get:
      consumes:
      - application/json
      description: 分页获取合约代码相关交易
      parameters:
      - description: 合约代码哈希
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 合约代码相关交易
      tags:
      - 合约代码
  /contractcodess/{name}:
    get:
      consumes:
      - application/json
      description: 获取合约代码信息
      parameters:
      - description: 合约代码哈希
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 合约代码信息
      tags:
      - 合约代码
  /contracts:
    get:
      consumes:
      - application/json
      description: 分页获取部署合约列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 部署合约列表
      tags:
      - 部署合约
  /contracts/{name}:
    get:
      consumes:
      - application/json
      description: 获取部署合约信息
      parameters:
      - description: 合约名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 部署合约信息
      tags:
      - 部署合约
  /contracts/{name}/transactions:
    get:
      consumes:
      - application/json
      description: 分页获取部署合约相关交易
      parameters:
      - description: 合约名称
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 部署合约相关交易
      tags:
      - 部署合约
...
  /search/{name}:
    get:
      consumes:
      - application/json
      description: 搜索区块高度、区块哈希、交易哈希、账户地址、资产、合约、合约代码
      parameters:
      - description: 未知
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 搜索
      tags:
      - 区块
  /transactions:
    get:
      consumes:
      - application/json
      description: 分页获取交易列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 交易列表
      tags:
      - 交易
...
swagger: "2.0"